fim do crud

+++ src/main/kotlin/br/com/wa/api/repository/UserRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UserRepository : JpaRepository<UserEntity, Int>
+interface UserRepository : JpaRepository<UserEntity, Long>
